requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
# sse_starlette
# bs4
# pypdf
# faiss-cpu
# faiss-gpu
wikipedia
arxiv